+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [#40](https://github.com/creecros/Mailmagik/issues/40) Send notification for incoming task emails. The mailbox is scanned by the regular mailmagik:fetchmail cron job.
 
+- [#41](https://github.com/creecros/Mailmagik/issues/41) Config option to select the server encoding, for compatibility with Exchange IMAP.
+
 - [#42](https://github.com/creecros/Mailmagik/issues/42) Subject parsing: Allow multi-word attribute values for columns, categories and tags, by quoting them with any of "", '', ‘’, “”, «» and „“.
 
 ### Changed